move footer to folder
