+++ src/app/_layout/footer/assets/index.ts
